Template/requirements.txt
flask
requests
google-generativeai
python-dotenv
gunicorn